fix: rename namespaces in tests folder
